requirements.txt
absl-py @ file:///C:/ci/absl-py_1623867338309/work
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
autopy==4.0.0
certifi==2021.10.8
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
fonttools==4.25.0
kiwisolver @ file:///C:/ci/kiwisolver_1612282618948/work
matplotlib @ file:///C:/ci/matplotlib-suite_1638289827560/work
mediapipe==0.8.9
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
MouseInfo==0.1.3
munkres==1.1.4
numpy @ file:///C:/ci/numpy_and_numpy_base_1634106873805/work
olefile==0.46
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
Pillow==8.4.0
protobuf==3.19.1
psutil @ file:///C:/ci/psutil_1612298033174/work
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput @ file:///D:/python/pycharm%20project/hand_last/gesture-control/pynput-1.4.5-py2.py3-none-any.whl
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQtChart==5.15.5
PyQtChart-Qt5==5.15.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytweening==1.0.4
pywin32==302
six @ file:///tmp/build/80754af9/six_1623709665295/work
torch==1.10.0
torchvision==0.11.1
tornado @ file:///C:/ci/tornado_1606935947090/work
typing_extensions==4.0.1
wincertstore==0.2
